Move AnalysisPluginFactory
