+++ src/main/java/org/harmony_analyser/application/HarmonyAnalyser.java
@@ -4,7 +4,6 @@ import org.harmony_analyser.jharmonyanalyser.services.*;
 import org.harmony_analyser.application.visualizations.*;
 import org.harmony_analyser.jharmonyanalyser.chord_analyser.*;
 import org.harmony_analyser.jharmonyanalyser.chroma_analyser.Chroma;
-import org.harmony_analyser.jharmonyanalyser.plugins.*;
 
 import javax.sound.midi.*;
 import javax.swing.*;
